Remove privoxy
