--- shadowsocks-csharp/Data/privoxy_conf.txt
@@ -1,5 +1,0 @@
-listen-address __PRIVOXY_BIND_IP__:__PRIVOXY_BIND_PORT__
-show-on-task-bar 0
-activity-animation 0
-forward-socks5 / 127.0.0.1:__SOCKS_PORT__ .
-hide-console
